requirements.txt
requests
beautifulsoup4
streamlit
numpy
Pillow
flask
keras
tensorflow
